Functions from project settings extracted as utils
