+++ src/app/projects/components/project-settings/components/data-schema/data-schema.component.ts
@@ -19,7 +19,7 @@ export class DataSchemaComponent implements OnInit {
   @Input() pKeyValid: boolean;
   @Input() dataHandlerHelper: DataHandlerHelper;
 
-  tooltipsArray: string[] = [];
+  tooltipsArray: string[];
   attributeVisibilityStates = attributeVisibilityStates;
   somethingLoading: boolean = false;
