+++ src/app/util/download-helper-functions.ts
@@ -1,0 +1,40 @@
+
+export function downloadBlob(byteData: any, filename = 'file.zip') {
+    const blob = new Blob([byteData], {
+        type: "application/octet-stream"
+    })
+    const blobUrl = URL.createObjectURL(blob);
+
+    // Create a link element
+    const link = document.createElement("a");
+    link.href = blobUrl;
+    link.download = filename;
+    document.body.appendChild(link);
+    link.dispatchEvent(
+        new MouseEvent('click', {
+            bubbles: true,
+            cancelable: true,
+            view: window
+        })
+    );
+
+    document.body.removeChild(link);
+}
+
+export function downloadText(filename, text) {
+    if (!text) return;
+    const element = document.createElement('a');
+
+    element.setAttribute(
+        'href',
+        'data:text/plain;charset=utf-8,' + encodeURIComponent(text)
+    );
+    element.setAttribute('download', filename);
+
+    element.style.display = 'none';
+    document.body.appendChild(element);
+
+    element.click();
+
+    document.body.removeChild(element);
+}
